pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.10"
name = "purchase_predict"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    "ipython>=8.10",
    "jupyterlab>=3.0",
    "notebook",
    "kedro~=1.2.0",
    "scikit-learn>=1.7.2",
    "pandas-stubs>=2.3.3.260113",
    "kedro-datasets>=9.2.0",
    "kedro-viz>=12.3.0",
    "lightgbm>=4.6.0",
    "hyperopt>=0.2.7",
    "setuptools<70",
    "google-cloud-storage>=3.9.0",
    "pre-commit>=4.5.1",
]

[project.scripts]
"purchase-predict" = "purchase_predict.__main__:main"

[project.entry-points."kedro.hooks"]

[project.optional-dependencies]
docs = [
    "docutils<0.21",
    "sphinx>=5.3,<7.3",
    "sphinx_rtd_theme==2.0.0",
    "nbsphinx==0.8.1",
    "sphinx-autodoc-typehints==1.20.2",
    "sphinx_copybutton==0.5.2",
    "ipykernel>=5.3, <7.0",
    "Jinja2<3.2.0",
    "myst-parser>=1.0,<2.1"
]
dev = [
    "pytest-cov>=3,<7",
    "pytest-mock>=1.7.1, <2.0",
    "pytest~=7.2",
    "ruff~=0.12.0"
]

[tool.setuptools.dynamic]
version = {attr = "purchase_predict.__version__"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.kedro]
package_name = "purchase_predict"
project_name = "purchase-predict"
kedro_init_version = "1.2.0"
tools = "['Linting', 'Testing', 'Custom Logging', 'Documentation', 'Data Structure']"
example_pipeline = "False"
source_dir = "src"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/purchase_predict -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff]
line-length = 120
show-fixes = true

[tool.ruff.lint]
select = [
    "F",   # Pyflakes
    "W",   # pycodestyle
    "E",   # pycodestyle
    "I",   # isort
    "UP",  # pyupgrade
    "PL",  # Pylint
    "T201", # Print Statement
]
ignore = ["E501"]  # Ruff format takes care of line-too-long

[tool.kedro_telemetry]
project_id = "2b7f871bf26e49d3bac90d2af9f5a352"

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]
